♻️ refactor(ai): extract AI processing logic into separate helper module
The commit consolidates duplicate AI processing logic into a new aiHelpers.ts
module, improving code maintainability and reducing redundancy across different
event processing methods.
